# openapi.yaml
openapi: 3.0.3
info:
  title: Vortex API
  description: Vortex 即时通讯系统 API 文档
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
  - url: https://api.example.com
    description: 生产服务器
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token认证
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: invalid_input
    User:
      type: object
      properties:
        username:
          type: string
          example: johndoe
        email:
          type: string
          example: john@example.com
        public_id:
          type: string
          example: abc123XYZ
    Message:
      type: object
      properties:
        id:
          type: string
          example: msg_xyz789
        conv_id:
          type: string
          example: conv_abc123
        sender_id:
          type: string
          example: abc123XYZ
        content:
          type: string
          example: Hello!
        content_type:
          type: string
          example: text
        created_at:
          type: string
          format: date-time
          example: "2026-05-14T12:00:00Z"
    Group:
      type: object
      properties:
        id:
          type: string
          example: grp_abc123
        name:
          type: string
          example: My Group
        description:
          type: string
          example: A group for friends
        owner_id:
          type: string
          example: abc123XYZ
        created_at:
          type: string
          format: date-time
    FriendRequest:
      type: object
      properties:
        id:
          type: integer
          example: 123
        status:
          type: string
          enum: [pending, accepted, rejected]
        sender_public_id:
          type: string
        receiver_public_id:
          type: string
paths:
...
  /api/auth/register:
    post:
      tags: [auth]
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 20
                  example: johndoe
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
                  example: SecurePass123!
                email:
                  type: string
                  format: email
                  example: john@example.com
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: 请求参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 用户名已存在
  /api/auth/login:
    post:
      tags: [auth]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: johndoe
                password:
                  type: string
                  example: SecurePass123!
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: 用户名或密码错误
        '429':
          description: 登录失败次数过多
  /api/auth/me:
    get:
      tags: [auth]
      summary: 获取当前用户信息
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: 未授权
  /api/auth/logout:
    post:
      tags: [auth]
      summary: 用户登出
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: logged out successfully
  /api/auth/{publicId}:
    put:
      tags: [auth]
      summary: 更新用户信息
      security:
        - bearerAuth: []
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
    delete:
      tags: [auth]
      summary: 删除用户
      security:
        - bearerAuth: []
      parameters:
        - name: publicId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
        '403':
          description: 无权限
  /api/messages/send:
    post:
      tags: [messages]
      summary: 发送消息
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conv_id, content]
              properties:
                conv_id:
                  type: string
                  example: conv_abc123
                content:
                  type: string
                  maxLength: 1000
                  example: Hello!
                text:
                  type: string
                  maxLength: 1000
                content_type:
                  type: string
                  example: text
                client_msg_id:
                  type: string
                  description: 客户端消息ID，用于幂等性
      responses:
        '201':
          description: 发送成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    $ref: '#/components/schemas/Message'
        '429':
          description: 发送过于频繁
  /api/messages:
    get:
      tags: [messages]
      summary: 获取消息列表
      security:
        - bearerAuth: []
      parameters:
        - name: conv_id
          in: query
          required: true
          schema:
            type: string
        - name: lastMsgId
          in: query
          schema:
            type: integer
            description: 游标模式，只返回 msg_id > lastMsgId 的消息（与 page 互斥）
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  page:
                    type: integer
                  page_size:
                    type: integer
                  has_more:
                    type: boolean
                  last_msg_id:
                    type: integer
                    description: 仅游标模式返回
  /api/messages/recall/{msgId}:
    post:
      tags: [messages]
      summary: 撤回消息
      security:
        - bearerAuth: []
      parameters:
        - name: msgId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 撤回成功
        '400':
          description: 超过撤回时间限制
        '403':
          description: 只能撤回自己的消息
  /api/check:
    get:
      tags: [messages]
      summary: 检查新消息
      security:
        - bearerAuth: []
      parameters:
        - name: lastMsgId
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 1
                    description: "0=无事发生, 1=新消息, 2=好友/群组更新, 3=两者都有"
                  updated:
                    type: array
                    items:
                      type: string
                    example: ["p_abc123_def456", "g_xyz789"]
  /api/conversations:
    get:
      tags: [messages]
      summary: 获取会话列表
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      type: object
                      properties:
                        conv_id:
                          type: string
                        type:
                          type: string
                        name:
                          type: string
                        unread_count:
                          type: integer
                  total:
                    type: integer
  /api/friends/request/send/{targetPublicId}:
    post:
      tags: [friends]
      summary: 发送好友请求
      security:
        - bearerAuth: []
      parameters:
        - name: targetPublicId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: 发送成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FriendRequest'
        '409':
          description: 已发送过请求或已是好友
  /api/friends/requests:
    get:
      tags: [friends]
      summary: 获取好友请求列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  sent:
                    type: array
                    items:
                      $ref: '#/components/schemas/FriendRequest'
                  received:
                    type: array
                    items:
                      $ref: '#/components/schemas/FriendRequest'
  /api/friends/request/{requestId}/accept:
    post:
      tags: [friends]
      summary: 接受好友请求
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 接受成功
  /api/friends/request/{requestId}/reject:
    post:
      tags: [friends]
      summary: 拒绝好友请求
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 拒绝成功
  /api/friends/request/{requestId}:
    delete:
      tags: [friends]
      summary: 取消好友请求
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 取消成功
  /api/blocks/{targetPublicId}:
    post:
      tags: [friends]
      summary: 拉黑用户
      security:
        - bearerAuth: []
      parameters:
        - name: targetPublicId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 拉黑成功
    delete:
      tags: [friends]
      summary: 取消拉黑
      security:
        - bearerAuth: []
      parameters:
        - name: targetPublicId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 取消拉黑成功
  /api/groups:
    post:
      tags: [groups]
      summary: 创建群组
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 50
                  example: My Group
                description:
                  type: string
                  example: A group for friends
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  group:
                    $ref: '#/components/schemas/Group'
  /api/groups/{id}:
    get:
      tags: [groups]
      summary: 获取群组信息
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  group:
                    $ref: '#/components/schemas/Group'
    put:
      tags: [groups]
      summary: 更新群组
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '200':
          description: 更新成功
    delete:
      tags: [groups]
      summary: 删除群组
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
  /api/groups/{id}/join:
    post:
      tags: [groups]
      summary: 加入群组
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 加入成功
  /api/groups/{id}/leave:
    post:
      tags: [groups]
      summary: 退出群组
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 退出成功
  /api/groups/{id}/members/{memberPublicId}:
    delete:
      tags: [groups]
      summary: 踢出成员
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: memberPublicId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 踢出成功
        '403':
          description: 不是群主
  /api/messages/{msgId}:
    get:
      tags: [messages]
      summary: 获取消息详情
      security:
        - bearerAuth: []
      parameters:
        - name: msgId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg_id:
                    type: integer
                  conv_id:
                    type: string
                  from_uid:
                    type: integer
                  content:
                    type: string
                  ts:
                    type: integer
                  is_recalled:
                    type: boolean
        '404':
          description: 消息不存在
  /api/conversations/count:
    get:
      tags: [messages]
      summary: 获取用户会话数
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                  count:
                    type: integer
  /api/conversations/{convId}/participants:
    get:
      tags: [messages]
      summary: 获取会话参与者
      security:
        - bearerAuth: []
      parameters:
        - name: convId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  conv_id:
                    type: string
                  participants:
                    type: array
                    items:
                      type: integer
        '404':
          description: 会话不存在
  /api/conversations/{convId}/blocked/{userId}:
    get:
      tags: [messages]
      summary: 检查屏蔽状态
      security:
        - bearerAuth: []
      parameters:
        - name: convId
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  conv_id:
                    type: string
                  user_id:
                    type: integer
                  is_blocked:
                    type: boolean
  /api/friends/requests/pending:
    get:
      tags: [friends]
      summary: 获取待处理好友请求
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        sender_id:
                          type: integer
                        receiver_id:
                          type: integer
                        status:
                          type: string
                          enum: [pending, accepted, rejected]
                        ts:
                          type: integer
  /api/groups/{id}/members/count:
    get:
      tags: [groups]
      summary: 获取群组成员数
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  group_id:
                    type: string
                  count:
                    type: integer
        '404':
          description: 群组不存在
  /api/files/presign:
    post:
      tags: [files]
      summary: 获取预签名URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operation]
              properties:
                operation:
                  type: string
                  enum: [upload, download]
                  example: upload
                conv_id:
                  type: string
                  description: 上传时必填
                file_ext:
                  type: string
                  description: 上传时必填，文件扩展名
                file_key:
                  type: string
                  description: 下载时必填，文件Key
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                  file_key:
                    type: string
                  expires_in:
                    type: integer
        '503':
          description: S3服务未配置